Move GetBytesPerCharacter to ProtocolUtility.
